Added Different Data Structure
Created new Classes to facilitate the Piece superclass, that is extended
by Rook, Knight, and Bishop.
Created a seperate PiecePositon Class that would be able to work with
user input easier as the user will initially input a row/col value i.e.
A1 to select and move the Pieces